fix: polymorphic type bounds
This signature change has been introduced in Shapeless 3.0.4. Unfortunately, Scala 3.3 compiler did not regard `[t <: X] => F[t]` as a subtype of `[t] => F[t]` (I am not sure if the compiler rejects this subtyping relation for all X and F), so we need to explicitly write the bounds down.