Fix an illegal match type case under SIP-56.
`CodecBinding` per se is not a valid type constructor for match
type cases under SIP-56.

We manually expand the `Tuple.InverseMap` so that we get a more
precise decomposition, which allows to write a match type case
that is valid under the new rules, without changing the behavior.

This change allows to compile `protocol_core` but not yet its tests.