Turn the abstract `type Lock[X]` into a `sealed trait Lock[X]`.
An abstract type like `Lock[X]` cannot be proven disjoint from
anything under SIP-56, since it *could* be instantiated to anything.
In particular, it could be instantiated to `type Lock[X] = Int`,
and then clearly even `Lock[Int]` and `Lock[String]` would not
be disjoint.

By making it a `sealed trait` instead, it is a concrete class type,
and so `Lock[Int]` and `Lock[String]` are actually disjoint.