Remove some tests that would still not pass under SIP-56.
No matter what we do, `Int` cannot be proven disjoint from `Any`
or from `1` (to take the first two tests that are removed). It
happened to work under the unspecified match types before SIP-56,
but there was evidence to guarantee its soundness.

It does not seem like the larger project needs those particular
tests to pass to be meaningful. So we suggest to remove them, and
focus on types that are actually provably disjoint.